Fixes bug where GA would never find an improvement from a mutation.
The problem was that, when comparing the fitness of a parent and child,
the numbers being compared were ALWAYS 0. Thus, since 0 is not greater
than 0, the result was always not an improvement. Now, when the GA runs
in its current state, the correct fitness values are compared each time.
After 100 mutations, the image that is shown at the end will PROBABLY
have a higher fitness than the initial. Since it's the last mutation
done, it's impossible to be sure if it was actually an improvement from
our most fit iteration, but there likely will have been at least a few
improvements up to that point.