Adds more structure to the TriangleMutation class
TriangleMutation is now instantiated with an added integer, its
stepValue. This value is what we will add or subtract from the gene
value. The GA will be responsible for telling the TriangleMutation what
to alter itself by every time.

Also, mutating an RGBA value is more sophisticated. TriangleMutation now
doesn't try to increase or decrease passed 255 or 0, respectively.