Fixes some bugs in the GA and alters the class to work with the new TriangleMutation class.
There was a bug in the GA where it would be using wrong values when
trying to repeat a mutation that previously resulted in an improvement.
The GA now properly climbs a single hill before checking another. Before
this, the algorithm was just effectively jumping around from hill to
hill.

The GA is updated to work with the new TriangleMutation class. As of
now, the added stepValue parameter in the constructor does not change
how the GA behaves. Prior to this change, every mutation would only
increase or decrease a value by 1. Right now, the GA only instantiates a
TriangleMutation class with strictly the value of 1. Thus, logically, no
change. Further testing will progressively alter this stepSize.