Replaces MainApplication class with the Application package.
The NewMain.java class acts as a sort of the CPU for the whole
application. The controllers handle the GUI and the GA handles the
actual pathfinding algorithm. NewMain contains an AnimationTimer which
acts as the application loop. Here, main tells the GA when to continue
iterating through the mutations. When the GA finds images with better
fitness, it gives this information to NewMain. Every half a second, main
renders the most fit image and displays the value by calling the update
method in mainController. When buttons are pressed on the GUI, the
mainController class calls up main to change what it needs to change.
For example: Stop button pressed, controller tells main, which stops the
loop from running and stops the GA.