Optimizes Hard Mutation methods and implements a ComboBox GUI to select which mutation to perform.
There is a Hard Mutation and a Soft Mutation, where hard is the one
recently implemented and soft is the original mutation. The GUI
basically tells GA whether IS_HARD_MUTATE_MODE is true or false.

Removes the checkIfHardMutationWasImprovement method as it was a
duplicate to the original check improvement method. Also, only one
integer is initialized now to handle the new gene value during a
hardMutate() call, as opposed to 10 different ones. Finally, now that
we're using only a single integer, the unnecessary second switch
statement is also removed.