Updates GUI and adds many bugfixes to GA.
There were a lot of things that weren't working correctly once the GA
moved to a multithreaded format. The biggest issue was both GAs
receiving the same "initial fitness." This made it extremely hard for
the GA with the lower initial fitness to ever start making improvements,
as the first improvement it was tasked to do would have to be an
improvement from the initial population with the higher fitness. For
example, if the lower population had a fitness of 40 and the higher
population had a fitness of 70, the lower population would have to first
find an improvement that would give it a value of over 70.. it basically
made it useless. Now, they both run as they should.

The threads now properly pause and start when given a start/stop command
from the GUI.

The ordered list of genomes created in InitialPopulations now gets
updated for the fitness values if the application decides to go with a
white background instead of a black. Before, the values would not change
to the (improved) values when the application would decide to use a
white background.

Genomes now have a copy of its Image.

There are a lot of things added to the gui, but anything with an N/A at
the end of it outside of the elapsed time is currently not being
calculated. You can now switch from the initial highest fit genome from
tribe 0, the initial highest fit genome from tribe 1, or just let the
application show you the most fit of the two.

The print all genome fitness button prints the INITIAL values for the
fitness of all the genomes. These are the values that were are created
in InitialPopulation. These values do NOT currently get updated.