Gets rid of old comments and simplifies code with how some things in the GUI are handled.
Before, when trying to cycle through the different top genome from each
tribe, we would handle it differently if the GA had already run once or
if it hadn't been started yet. A previous bug, which sometimes gave a GA
an incorrect current best fitness, made this necessary. Now we can use
the same code for both instances.

Also, now the threads get put to sleep when the application is paused.
Before, we would have them print empty lines to the console so they
wouldn't die.