Fixes bug where the GUI would display a wrong fitness. (I think)
I think the problem was the GUI would sometimes display the incorrect
fitness. It wasn't that it was an old best fitness or it was displaying
an older picture. I think the genome was always correct only when
calling getFit() from the GA, the value that was returned was the wrong
value. I suspect this problem has probably been here for a while but
didn't get noticed. Oops.