Makes GUI more user friendly and now updates the genome in the list of tribes to show the appropriate genome when the user selects specific genomes.
The GUI now has buttons disabled at startup to guide the user to first
upload an image then initialize the population before allowing the user
to try and start the GA (and also break everything). Also selecting
specific genomes is more intuitive now with the dimming combo box for
selecting genomes. When it's not dimmed is when it actually makes a
difference.

Also, speaking of displaying user selected genomes, before when the
program would attempt to show this, every genome was always from the
initial population. Considering only one from each tribe is ever
changing right now, this was only a problem for 2 out of 200.
Regardless, the initial genome from each tribe is now the updated
version, as opposed to the initial.