Fixes a couple bugs when closing the application and when trying to draw a user selected genome.
When closing the main window the application would still be running
thanks to the worker threads not being properly closed. This update adds
an override to the stop() method from the Application class found in
NewMain to make sure the threads stop. Not sure if this is the most
sophisticated way of stopping threads but it works in our case.

This update also gives the NewMain class it's own
TriangleGenome.renderer object. By using this when the application needs
to display a user selected genome, there are no problems displaying the
correct genome while the pathfinder is running. The heap space used by
the program in this update saw virtually no change from the previously
pushed version that was made for optimization.