Main now takes a snapshot of the statistics every 5 seconds. GUI can now also save all the statistical snapshots or the currently displayed genome to a text file.
A genome gets saved into a 200 line text file, where each line
represents each triangle and each of the 10 variables are seperated with
a space.

The snapshot saves all the statistics saved to the gui to a single
string, with each value seperated by a space. This string begins with
the current elapsed time. The strings are then placed in an
ArrayList<String>, where the first item in the list is the first line,
the second item is the second line, etc.