Application now starts up with a default target image, number of tribes, starting population, and selected genome, but leads to some unwanted results. Also fixes a minor bug.
Consequently the only way to change the number of tribes is to edit the
main method now that a default number of tribes is selected. Will be
fixed soon just maybe not tonight.

Also there was a bug where the wrong genomes for the user to select were
displayed. It was hardcoded to always show 100 instead of taking into
account the actual population. Now it properly shows the correct amount
based on the actual population of the tribe.