Adds an example excel spreadsheet with data and graph
This is data from two runs that ran for approximately 3 hours and 30 minutes. The difference between the two runs are that one used exclusively hillclimbing while the other used the current version of the GA with crossover. This is intended to be strictly for example purposes to show how relatively easy it is to use the data outputted by the application to create graphs. This data may be useful for the final graphs but I'm unsure.